Hopefully source is self-contained
Hopefully making it so that you can just build from by unzipping the source code and having it work from the box.

This was done by making cmake look for SFML in the root directory and then just putting the necessary SFML version in that folder. In this case, the 2.5.1 32-bit version.